--- Website/src/actions/trainingActions.js
@@ -1,5 +1,0 @@
-export const INITIALIZE_TRAINING_HISTORY = 'INITIALIZE_TRAINING_HISTORY';
-export const initializeTrainingHistory = history => ({
-  type: INITIALIZE_TRAINING_HISTORY,
-  payload: history
-});
