+++ Website/src/components/app/PageContainer/Header/ControlledNavDrawer/NavDrawer/DrawerNavList/DrawerNavList.js
@@ -8,32 +8,16 @@ import './DrawerNavList.scss';
 function DrawerNavList({ routes, handleClose }) {
   return (
     <List className="drawer-nav-list">
-      {routes.map(({ to, text, nestedRoutes }) => (
+      {routes.map(({ to, text }) => (
         <li key={to}>
           <NavLink
             to={to}
             onClick={handleClose}
-            className="drawer-nav-list__link drawer-nav-list__general-link"
+            className="drawer-nav-list__link"
             activeClassName="active"
           >
             {text}
           </NavLink>
-          {nestedRoutes && (
-            <List>
-              {nestedRoutes.map(({ to, text }) => (
-                <li key={to}>
-                  <NavLink
-                    to={to}
-                    onClick={handleClose}
-                    className="drawer-nav-list__link drawer-nav-list__nested-link"
-                    activeClassName="active"
-                  >
-                    {text}
-                  </NavLink>
-                </li>
-              ))}
-            </List>
-          )}
         </li>
       ))}
     </List>
@@ -44,13 +28,7 @@ DrawerNavList.propTypes = {
   routes: PropTypes.arrayOf(
     PropTypes.shape({
       to: PropTypes.string.isRequired,
-      text: PropTypes.string.isRequired,
-      nestedRoutes: PropTypes.arrayOf(
-        PropTypes.shape({
-          to: PropTypes.string.isRequired,
-          text: PropTypes.string.isRequired
-        }).isRequired
-      )
+      text: PropTypes.string.isRequired
     }).isRequired
   ).isRequired,
   handleClose: PropTypes.func.isRequired
