+++ Website/src/components/app/PageContainer/Header/HeaderNavList/HeaderNavList.scss
@@ -1,16 +1,22 @@
+@import 'src/theme/mixins';
+
 .header-nav-list {
   margin: 0;
   display: flex;
-  justify-content: space-around;
   padding: 0;
   list-style: none;
-  width: 100%;
-  height: 100%;
+
+  @include theme-breakpoints-down(md) {
+    display: none;
+  }
 
   &__item {
+    margin: 0 16px;
     display: flex;
-    justify-content: center;
-    height: 100%;
+
+    @include theme-breakpoints-down(lg) {
+      margin: 0 1vw;
+    }
   }
 
   &__link {
@@ -19,7 +25,6 @@
     align-items: center;
     padding: 0 14px;
     font-size: 1rem;
-    height: 100%;
     text-decoration: none;
     color: inherit;
     outline: 0;
@@ -39,8 +44,8 @@
       content: '';
       position: absolute;
       bottom: 0;
-      left: 0;
-      right: 0;
+      left: 14px;
+      right: 14px;
       height: 3px;
       background-color: var(--palette-primary-contrast-text);
       transform: scaleY(0);
