+++ Website/src/components/app/PageContainer/PageContainer.scss
@@ -3,33 +3,12 @@
 .page-container {
   display: flex;
 
-  &__side-nav {
-    position: fixed;
-    top: 0;
-    bottom: 0;
-    overflow-y: auto;
-    width: 240px;
-    padding-top: 88px;
-    border-right: 1px solid var(--palette-divider);
-    background-color: var(--palette-background-paper);
-
-    @include theme-breakpoints-down(lg) {
-      display: none;
-    }
-  }
-
   &__main {
-    margin-top: 64px;
+    margin-top: 72px;
     flex-grow: 1;
 
-    @include theme-breakpoints-down(sm) {
+    @include theme-breakpoints-down(md) {
       margin-top: 56px;
-    }
-
-    &--has-nav {
-      @include theme-breakpoints-up(lg) {
-        padding-left: 240px;
-      }
     }
   }
 }
