--- Website/src/components/app/PageContainer/SideNavList/SideNavList.scss
@@ -1,37 +1,0 @@
-@import 'src/theme/mixins';
-
-.side-nav-list {
-  width: 100%;
-
-  &__link {
-    position: relative;
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    min-height: 40px;
-    padding: 0 24px 0;
-    cursor: pointer;
-    user-select: none;
-    outline: 0;
-    text-decoration: none;
-
-    @include theme-typography(body2);
-
-    color: var(--palette-text-primary);
-    transition: background-color var(--transition-duration-shortest)
-      var(--transition-easing-ease-in-out);
-
-    &:hover,
-    &:focus {
-      background-color: var(--palette-action-hover);
-    }
-
-    &:active {
-      background-color: var(--palette-action-selected);
-    }
-
-    &.active {
-      font-weight: 500;
-    }
-  }
-}
