--- Website/src/components/app/PageContainer/SideNavList/index.js
@@ -1,1 +1,0 @@
-export { default } from './SideNavList';
