+++ Website/src/components/routes/WordSetPairs/WordSetPairsToolbar/WordSetPairsToolbar.scss
@@ -2,12 +2,12 @@
 
 .word-set-pairs-toolbar {
   position: sticky;
-  top: 64px;
+  top: 72px;
   background-color: var(--palette-background-paper);
   z-index: 1;
   border-bottom: 1px solid var(--palette-divider);
 
-  @include theme-breakpoints-down(sm) {
+  @include theme-breakpoints-down(md) {
     top: 56px;
   }
 }
