+++ Website/src/reducers/index.js
@@ -5,7 +5,6 @@ import auth from './auth';
 import profile from './profile';
 import dictionary from './dictionary';
 import wordSet from './wordSet';
-import trainingHistory from './trainingHistory';
 
 const rootReducer = (state, action) => {
   if (action.type === DENY_ACCESS) {
@@ -17,8 +16,7 @@ const rootReducer = (state, action) => {
     auth,
     profile,
     dictionary,
-    wordSet,
-    trainingHistory
+    wordSet
   });
 
   return appReducer(state, action);
